+++ src/features/XIT/BURN/PlanetHeader.vue
@@ -4,6 +4,7 @@ import { showBuffer } from '@src/infrastructure/prun-ui/buffers';
 import PrunButton from '@src/components/PrunButton.vue';
 import { PlanetBurn } from '@src/core/burn';
 import { countDays } from '@src/features/XIT/BURN/utils';
+import { planetContextMenu } from '@src/components/planet-context-menu/planet-context-menu';
 import { useTileState } from '@src/features/XIT/BURN/tile-state';
 
 const { burn } = defineProps<{
@@ -20,7 +21,11 @@ const nameColspan = computed(() => (io.value ? 6 : 4));
 
 <template>
   <tr :class="$style.row">
-    <td :colspan="nameColspan" :class="$style.cell" @click="onClick">
+    <td
+      :colspan="nameColspan"
+      :class="$style.cell"
+      @click="onClick"
+      @contextmenu.prevent="burn.naturalId && planetContextMenu.showMenu($event, burn.naturalId)">
       <span v-if="hasMinimize" :class="$style.minimize">
         {{ minimized ? '+' : '-' }}
       </span>